Fixed #1.
An optimization made to cut down on queries to the skin server made it so if you attempted to join a server before you started a single-player world, you would have no skin. This was because it now only attempted to query the skin server on single-player world joins. We now query it again when connecting to a server.